move virt manager settings from db to config file
This means we only support one (optional) RHEV-M instance, instead of
many. It makes things a lot simpler and lets us fix the issue with
leaking VMs in beakerd.virt_recipes if the db transaction rolls back.

+++ Server/bkr/server/model.py
@@ -51,7 +51,6 @@ import netaddr
 from bkr.common.helpers import Flock, makedirs_ignore, unlink_ignore
 import subprocess
 from turbogears import identity
-from contextlib import contextmanager
 import ovirtsdk.api
 from datetime import timedelta, date, datetime
 from hashlib import md5
@@ -1092,8 +1091,6 @@ system_resource_table = Table('system_resource', metadata,
 virt_resource_table = Table('virt_resource', metadata,
     Column('id', Integer, ForeignKey('recipe_resource.id',
             name='virt_resource_id_fk'), primary_key=True),
-    Column('manager_id', Integer, ForeignKey('virt_manager.id',
-            name='virt_resource_manager_id_fk'), nullable=False),
     Column('system_name', Unicode(2048), nullable=False),
     Column('lab_controller_id', Integer, ForeignKey('lab_controller.id',
             name='virt_resource_lab_controller_id_fk'), nullable=False),
@@ -1228,16 +1225,6 @@ rendered_kickstart_table = Table('rendered_kickstart', metadata,
     # it means fetch the kickstart from there instead
     Column('kickstart', UnicodeText),
     Column('url', UnicodeText),
-    mysql_engine='InnoDB',
-)
-
-virt_manager_table = Table('virt_manager', metadata,
-    Column('id', Integer, primary_key=True),
-    Column('url', Unicode(255), unique=True),
-    Column('username', Unicode(255)),
-    Column('password', Unicode(255)),
-    Column('disabled', Boolean, nullable=False, default=False, index=True),
-    Column('removed', DateTime, nullable=True, default=None, index=True),
     mysql_engine='InnoDB',
 )
 
@@ -5200,9 +5187,10 @@ class Recipe(TaskBase):
                     field_name=u'Distro Tree', old_value=u'',
                     new_value=unicode(self.distro_tree)))
         elif isinstance(self.resource, VirtResource):
-            self.resource.manager.start_install(self.resource.system_name,
-                    self.distro_tree, install_options.kernel_options_str,
-                    self.resource.lab_controller)
+            with VirtManager() as manager:
+                manager.start_install(self.resource.system_name,
+                        self.distro_tree, install_options.kernel_options_str,
+                        self.resource.lab_controller)
             self.tasks[0].start()
 
     def cleanup(self):
@@ -5956,10 +5944,8 @@ class VirtResource(RecipeResource):
     a hypervisor attached to Beaker.
     """
 
-    def __init__(self, manager, system_name, lab_controller):
-        self.manager = manager
+    def __init__(self, system_name):
         self.system_name = system_name
-        self.lab_controller = lab_controller
 
     @property
     def link(self):
@@ -5969,13 +5955,21 @@ class VirtResource(RecipeResource):
         return global_install_options().combined_with(
                 distro_tree.install_options())
 
-    def allocate(self):
-        pass
+    def allocate(self, manager, lab_controllers):
+        log.debug('Creating vm for recipe %s', self.recipe.id)
+        self.lab_controller = manager.create_vm(self.system_name,
+                lab_controllers)
 
     def release(self):
-        log.debug('Releasing virt guest %s for recipe %s',
-                self.system_name, self.recipe.id)
-        self.manager.destroy_vm(self.system_name)
+        try:
+            log.debug('Releasing vm %s for recipe %s',
+                    self.system_name, self.recipe.id)
+            with VirtManager() as manager:
+                manager.destroy_vm(self.system_name)
+        except Exception, e:
+            log.exception('Failed to destroy vm %s, leaked!',
+                    self.system_name)
+            # suppress exception, nothing more we can do now
 
 
 class GuestResource(RecipeResource):
@@ -6393,81 +6387,74 @@ class ConfigValueInt(MappedObject):
         if valid_from:
             self.valid_from = valid_from
 
-class VirtManager(MappedObject):
+class VirtManager(object):
 
-    def __repr__(self):
-        return '%s(id=%r, url=%r)' % (self.__class__.__name__, self.id, self.url)
+    def __init__(self):
+        self.api = None
 
-    @contextmanager
-    def api(self):
-        api = ovirtsdk.api.API(url=self.url, username=self.username, password=self.password,
+    def __enter__(self):
+        self.api = ovirtsdk.api.API(url=get('ovirt.api_url'),
+                username=get('ovirt.username'), password=get('ovirt.password'),
                 # XXX add some means to specify SSL CA cert
                 insecure=True)
-        try:
-            yield api
-        finally:
-            api.disconnect()
+        return self
 
-    @classmethod
-    def create_vm_on_any(cls, *args, **kwargs):
-        for manager in cls.query.filter_by(disabled=False, removed=None):
-            try:
-                log.debug('Trying to create VM with args %r %r on %r', args, kwargs, manager)
-                return manager.create_vm(*args, **kwargs)
-            except VMCreationFailedException:
-                continue
-        raise VMCreationFailedException('Unable to create VM on any manager')
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        self.api, api = None, self.api
+        api.disconnect()
 
     def create_vm(self, name, lab_controllers):
+        if self.api is None:
+            raise RuntimeError('Context manager was not entered')
         from ovirtsdk.xml.params import VM, Template, NIC, Network, Disk, StorageDomains
         # Default of 1GB memory and 20GB disk
         memory = ConfigItem.by_name('default_guest_memory').current_value(1024) * 1024**2
         disk_size = ConfigItem.by_name('default_guest_disk_size').current_value(20) * 1024**3
         # Try to create the VM on every cluster that is in an acceptable data center
-        with self.api() as rh:
-            cluster_query = ' or '.join('datacenter.name=%s' % lc.data_center_name
-                    for lc in lab_controllers)
-            for cluster in rh.clusters.list(cluster_query):
-                vm = None
-                try:
-                    vm_definition = VM(name=name, memory=memory, cluster=cluster,
-                            type_='virtio26', template=Template(name='Blank'))
-                    vm = rh.vms.add(vm_definition)
-                    nic = NIC(name='eth0', interface='virtio', network=Network(name='rhevm'))
-                    vm.nics.add(nic)
-                    sd_query = ' or '.join('datacenter=%s' % lc.data_center_name
-                            for lc in lab_controllers)
-                    storage_domains = rh.storagedomains.list(sd_query)
-                    disk = Disk(storage_domains=StorageDomains(storage_domain=storage_domains),
-                            size=disk_size, type_='data', interface='virtio', format='cow',
-                            bootable=True)
-                    vm.disks.add(disk)
+        cluster_query = ' or '.join('datacenter.name=%s' % lc.data_center_name
+                for lc in lab_controllers)
+        for cluster in self.api.clusters.list(cluster_query):
+            log.debug('Trying to create vm %s on cluster %s', name, cluster.name)
+            vm = None
+            try:
+                vm_definition = VM(name=name, memory=memory, cluster=cluster,
+                        type_='virtio26', template=Template(name='Blank'))
+                vm = self.api.vms.add(vm_definition)
+                nic = NIC(name='eth0', interface='virtio', network=Network(name='rhevm'))
+                vm.nics.add(nic)
+                sd_query = ' or '.join('datacenter=%s' % lc.data_center_name
+                        for lc in lab_controllers)
+                storage_domains = self.api.storagedomains.list(sd_query)
+                disk = Disk(storage_domains=StorageDomains(storage_domain=storage_domains),
+                        size=disk_size, type_='data', interface='virtio', format='cow',
+                        bootable=True)
+                vm.disks.add(disk)
 
-                    # Wait up to twenty seconds(!) for the disk image to be created
-                    for _ in range(20):
-                        if rh.vms.get(name).status.state != 'image_locked':
-                            break
-                        time.sleep(1)
-                    state = rh.vms.get(name).status.state
-                    if state == 'image_locked':
-                        raise ValueError('VM %s state %s', name, state)
+                # Wait up to twenty seconds(!) for the disk image to be created
+                for _ in range(20):
+                    if self.api.vms.get(name).status.state != 'image_locked':
+                        break
+                    time.sleep(1)
+                state = self.api.vms.get(name).status.state
+                if state == 'image_locked':
+                    raise ValueError('VM %s state %s', name, state)
 
-                    dc_name = rh.datacenters.get(id=cluster.data_center.id).name
-                    return VirtResource(manager=self, system_name=name,
-                            lab_controller=LabController.by_data_center_name(dc_name))
-                except Exception:
-                    log.exception("Failed to create VM %r on %r cluster %r",
-                            name, self, cluster.name)
-                    if vm is not None:
-                        try:
-                            vm.delete()
-                        except Exception:
-                            pass
-                    continue
-            raise VMCreationFailedException(
-                    'No clusters on %r successfully created VM %s' % (self, name))
+                dc_name = self.api.datacenters.get(id=cluster.data_center.id).name
+                return LabController.by_data_center_name(dc_name)
+            except Exception:
+                log.exception("Failed to create VM %r on %r cluster %r",
+                        name, self, cluster.name)
+                if vm is not None:
+                    try:
+                        vm.delete()
+                    except Exception:
+                        pass
+                continue
+        raise VMCreationFailedException('No clusters successfully created VM %s' % name)
 
     def start_install(self, name, distro_tree, kernel_options, lab_controller):
+        if self.api is None:
+            raise RuntimeError('Context manager was not entered')
         from ovirtsdk.xml.params import OperatingSystem, Action, VM
         # RHEV can only handle a local path to kernel/initrd, so we rely on autofs for now :-(
         # XXX when this constraint is lifted, fix beakerd.virt_recipes too
@@ -6478,22 +6465,19 @@ class VirtManager(MappedObject):
         kernel_path = os.path.join(local_path, kernel.path)
         initrd_path = os.path.join(local_path, initrd.path)
         log.debug(u'Starting VM %s installing %s', name, distro_tree)
-        with self.api() as rh:
-            a = Action(vm=VM(os=OperatingSystem(kernel=kernel_path,
-                    initrd=initrd_path, cmdline=kernel_options)))
-            rh.vms.get(name).start(action=a)
+        a = Action(vm=VM(os=OperatingSystem(kernel=kernel_path,
+                initrd=initrd_path, cmdline=kernel_options)))
+        self.api.vms.get(name).start(action=a)
 
     def destroy_vm(self, name):
-        try:
-            with self.api() as rh:
-                vm = rh.vms.get(name)
-                if vm is not None:
-                    log.debug('Stopping %s on %r', name, self)
-                    vm.stop()
-                    log.debug('Deleting %s on %r', name, self)
-                    vm.delete()
-        except Exception, e:
-            log.exception(e) # and discard
+        if self.api is None:
+            raise RuntimeError('Context manager was not entered')
+        vm = self.api.vms.get(name)
+        if vm is not None:
+            log.debug('Stopping %s on %r', name, self)
+            vm.stop()
+            log.debug('Deleting %s on %r', name, self)
+            vm.delete()
 
 # set up mappers between identity tables and classes
 Hypervisor.mapper = mapper(Hypervisor, hypervisor_table)
@@ -6883,7 +6867,6 @@ mapper(SystemResource, system_resource_table, inherits=RecipeResource,
 mapper(VirtResource, virt_resource_table, inherits=RecipeResource,
         polymorphic_on=recipe_resource_table.c.type, polymorphic_identity=ResourceType.virt,
         properties={
-            'manager': relation(VirtManager),
             'lab_controller': relation(LabController),
         })
 mapper(GuestResource, guest_resource_table, inherits=RecipeResource,
@@ -6941,7 +6924,6 @@ mapper(ConfigValueString, config_value_string_table,
        properties = {'config_item': relation(ConfigItem, uselist=False),
                      'user': relation(User)}
       )
-mapper(VirtManager, virt_manager_table)
 
 
 ## Static list of device_classes -- used by master.kid
